installer: allow update only tomcat-public

+++ applications/installer/lib/upgrade-tomcat.sh
@@ -3,7 +3,12 @@ function upgrade_tomcat {
   tomcat_current_version=`cat /opt/tomcat-public/RELEASE-NOTES |grep "Apache Tomcat Version" | sed 's/^ *//;s/ *$//' |sed -E 's/^Apache Tomcat Version (.*)$/\1/'`
   tomcat_last_version=`curl -s 'https://tomcat.apache.org/download-80.cgi' |grep -Po '<h3 id="8\.5.*>(.*)</h3>'|sed -E 's/^<h3 id="8\.5.*>(.*)<\/h3>$/\1/'`
 
-  show_info_text "Upgrade tomcat-public (from version $tomcat_current_version to $tomcat_last_version)..."
+  show_info_text "Upgrade tomcats (from version $tomcat_current_version to $tomcat_last_version)..."
+
+  tomcat_local_exists=false
+  if [ -d "$DIR_APP_TOMCATLOCAL" ]; then
+    tomcat_local_exists=true
+  fi
 
   if [ "$tomcat_current_version" != "$tomcat_last_version" ]; then
     url_tomcat_base="https://archive.apache.org/dist/tomcat/tomcat-8/v$tomcat_last_version/bin"
@@ -45,8 +50,10 @@ function upgrade_tomcat {
     show_status " Backup old versions..."
     BACKUP_DIR_TOMCATPUBLIC=$(get_backup_dir $DIR_APP_TOMCATPUBLIC)
     mv $DIR_APP_TOMCATPUBLIC $BACKUP_DIR_TOMCATPUBLIC
-    BACKUP_DIR_TOMCATLOCAL=$(get_backup_dir $DIR_APP_TOMCATLOCAL)
-    mv $DIR_APP_TOMCATLOCAL $BACKUP_DIR_TOMCATLOCAL
+    if $tomcat_local_exists ; then
+      BACKUP_DIR_TOMCATLOCAL=$(get_backup_dir $DIR_APP_TOMCATLOCAL)
+      mv $DIR_APP_TOMCATLOCAL $BACKUP_DIR_TOMCATLOCAL
+    fi
     show_status_ok
 
     show_status " Deploy new version (tomcat-public)..."
@@ -58,14 +65,16 @@ function upgrade_tomcat {
     rm -rf "$DIR_APP_TOMCATPUBLIC"/webapps/*
     show_status_ok
 
-    show_status " Deploy new version (tomcat-local)..."
-    mkdir "$DIR_APP_TOMCATLOCAL"
-    tar xzf "$tomcat_app_file" -C "$DIR_APP_TOMCATLOCAL"
-    dir_into_tar=`dir $DIR_APP_TOMCATLOCAL | tr -d "\n"`
-    mv "$DIR_APP_TOMCATLOCAL"/"$dir_into_tar"/* "$DIR_APP_TOMCATLOCAL"
-    rm -rf "$DIR_APP_TOMCATLOCAL"/"$dir_into_tar"/
-    rm -rf "$DIR_APP_TOMCATLOCAL"/webapps/*
-    show_status_ok
+    if $tomcat_local_exists ; then
+      show_status " Deploy new version (tomcat-local)..."
+      mkdir "$DIR_APP_TOMCATLOCAL"
+      tar xzf "$tomcat_app_file" -C "$DIR_APP_TOMCATLOCAL"
+      dir_into_tar=`dir $DIR_APP_TOMCATLOCAL | tr -d "\n"`
+      mv "$DIR_APP_TOMCATLOCAL"/"$dir_into_tar"/* "$DIR_APP_TOMCATLOCAL"
+      rm -rf "$DIR_APP_TOMCATLOCAL"/"$dir_into_tar"/
+      rm -rf "$DIR_APP_TOMCATLOCAL"/webapps/*
+      show_status_ok
+    fi
 
     show_status " Configure (tomcat-public)..."
     cp "$BACKUP_DIR_TOMCATPUBLIC"/webapps/*.war "$DIR_APP_TOMCATPUBLIC"/webapps/
@@ -78,27 +87,33 @@ function upgrade_tomcat {
     chown tomcat-public.monet "$DIR_APP_TOMCATPUBLIC"/* -R
     show_status_ok
 
-    show_status " Configure (tomcat-local)..."
-    cp "$BACKUP_DIR_TOMCATLOCAL"/webapps/*.war "$DIR_APP_TOMCATLOCAL"/webapps/
-    cp -f "$BACKUP_DIR_TOMCATLOCAL"/bin/startup.sh "$DIR_APP_TOMCATLOCAL"/bin
-    cp -f "$BACKUP_DIR_TOMCATLOCAL"/bin/debugging.sh "$DIR_APP_TOMCATLOCAL"/bin
-    sed -i "s/8005/8006/g" "$DIR_APP_TOMCATLOCAL"/conf/server.xml
-    sed -i "s/8080/8081/g" "$DIR_APP_TOMCATLOCAL"/conf/server.xml
-    sed -i "s/8443/8444/g" "$DIR_APP_TOMCATLOCAL"/conf/server.xml
-    sed -i "s/8009/8010/g" "$DIR_APP_TOMCATLOCAL"/conf/server.xml
-    cp -f "$BACKUP_DIR_TOMCATLOCAL"/lib/log4j-1.2.17.jar "$DIR_APP_TOMCATLOCAL"/lib
-    cp -f "$BACKUP_DIR_TOMCATLOCAL"/lib/log4j.properties "$DIR_APP_TOMCATLOCAL"/lib
-    cp -f "$BACKUP_DIR_TOMCATLOCAL"/lib/vl-logging-1.1.43.jar "$DIR_APP_TOMCATLOCAL"/lib
-    chown tomcat-local.monet "$DIR_APP_TOMCATLOCAL"/* -R
-    show_status_ok
+    if $tomcat_local_exists ; then
+      show_status " Configure (tomcat-local)..."
+      cp "$BACKUP_DIR_TOMCATLOCAL"/webapps/*.war "$DIR_APP_TOMCATLOCAL"/webapps/
+      cp -f "$BACKUP_DIR_TOMCATLOCAL"/bin/startup.sh "$DIR_APP_TOMCATLOCAL"/bin
+      cp -f "$BACKUP_DIR_TOMCATLOCAL"/bin/debugging.sh "$DIR_APP_TOMCATLOCAL"/bin
+      sed -i "s/8005/8006/g" "$DIR_APP_TOMCATLOCAL"/conf/server.xml
+      sed -i "s/8080/8081/g" "$DIR_APP_TOMCATLOCAL"/conf/server.xml
+      sed -i "s/8443/8444/g" "$DIR_APP_TOMCATLOCAL"/conf/server.xml
+      sed -i "s/8009/8010/g" "$DIR_APP_TOMCATLOCAL"/conf/server.xml
+      cp -f "$BACKUP_DIR_TOMCATLOCAL"/lib/log4j-1.2.17.jar "$DIR_APP_TOMCATLOCAL"/lib
+      cp -f "$BACKUP_DIR_TOMCATLOCAL"/lib/log4j.properties "$DIR_APP_TOMCATLOCAL"/lib
+      cp -f "$BACKUP_DIR_TOMCATLOCAL"/lib/vl-logging-1.1.43.jar "$DIR_APP_TOMCATLOCAL"/lib
+      chown tomcat-local.monet "$DIR_APP_TOMCATLOCAL"/* -R
+      show_status_ok
+    fi
 
     show_status " Start services... "
     if [ "$OS_VERSION" = "6" ]; then
       /etc/init.d/tomcat-public start &> /dev/null
-      /etc/init.d/tomcat-local start &> /dev/null
+      if $tomcat_local_exists ; then
+        /etc/init.d/tomcat-local start &> /dev/null
+      fi
     else
       systemctl start tomcat-public
-      systemctl start tomcat-local
+      if $tomcat_local_exists ; then
+        systemctl start tomcat-local
+      fi
     fi
     show_status_ok
   else
